myPage연결

+++ src/main/java/com/izo/itaportal/controller/MyPageController.java
@@ -2,6 +2,7 @@ package com.izo.itaportal.controller;
 
 
 import com.izo.itaportal.model.Admin;
+import com.izo.itaportal.model.LoginResponse;
 import com.izo.itaportal.model.Professor;
 import com.izo.itaportal.model.Student;
 import com.izo.itaportal.service.MyPageService;
@@ -21,18 +22,21 @@ import javax.servlet.http.HttpSession;
 @RequestMapping("/user")
 public class MyPageController {
 
-
     @Autowired
     private MyPageService myPageService;
 
     @Autowired
     private HttpSession httpSession;
 
-    //학생  user/info?idUser=2024
-    //교수  user/info?idUser=1
-    //관리자 user/info?idUser=1234
     @GetMapping("/info")
-    public String myPage(Model model, Integer idUser) {
+    public String myPage(Model model) {
+        LoginResponse user = (LoginResponse) httpSession.getAttribute("loginUser");
+        if (user == null) {
+            // 세션에 사용자가 없으면 로그인 페이지로 리다이렉트
+            return "redirect:/login";
+        }
+        Integer idUser = user.getIdUser();
+
         if (myPageService.isStudent(idUser)) {
             model.addAttribute("user", myPageService.getStudentById(idUser));
             return "/myPage/stuMyPage";
@@ -48,42 +52,57 @@ public class MyPageController {
     }
 
     @PostMapping("/updateInfo")
-    public String updateInfo(Integer idUser, Student student, Professor professor, Admin admin) {
+    public String updateInfo(Student student, Professor professor, Admin admin) {
+        LoginResponse user = (LoginResponse) httpSession.getAttribute("loginUser");
+        if (user == null) {
+            // 세션에 사용자가 없으면 로그인 페이지로 리다이렉트
+            return "redirect:/login";
+        }
+        Integer idUser = user.getIdUser();
 
         if (myPageService.isStudent(idUser)) {
             myPageService.updateStudent(student);
-            return "redirect:/user/info?idUser=" + student.getIdUser();
+            return "redirect:/user/info";
         } else if (myPageService.isProfessor(idUser)) {
             myPageService.updateProfessor(professor);
-            return "redirect:/user/info?idUser=" + professor.getIdUser();
+            return "redirect:/user/info";
         } else if (myPageService.isAdmin(idUser)) {
             myPageService.updateAdmin(admin);
-            return "redirect:/user/info?idUser=" + admin.getIdUser();
+            return "redirect:/user/info";
         } else {
             return "/";
         }
     }
 
     @GetMapping("/deleteInfo")
-    public String deleteInfo(Integer idUser, Student student, Professor professor, Admin admin) {
+    public String deleteInfo() {
+        LoginResponse user = (LoginResponse) httpSession.getAttribute("loginUser");
+        if (user == null) {
+            // 세션에 사용자가 없으면 로그인 페이지로 리다이렉트
+            return "redirect:/login";
+        }
+        Integer idUser = user.getIdUser();
 
         if (myPageService.isStudent(idUser)) {
-
-            return "/myPage/deleteInfo?idUser=" + student.getIdUser();
+            return "/myPage/deleteInfo?idUser=" + idUser;
         } else if (myPageService.isProfessor(idUser)) {
-
-            return "/myPage/deleteInfo?idUser=" + professor.getIdUser();
+            return "/myPage/deleteInfo?idUser=" + idUser;
         } else if (myPageService.isAdmin(idUser)) {
-
-            return "/myPage/deleteInfo?idUser=" + admin.getIdUser();
+            return "/myPage/deleteInfo?idUser=" + idUser;
         } else {
             return "/";
         }
-
     }
 
     @PostMapping("/delete")
-    public String delete(Integer idUser) {
+    public String delete() {
+        LoginResponse user = (LoginResponse) httpSession.getAttribute("loginUser");
+        if (user == null) {
+            // 세션에 사용자가 없으면 로그인 페이지로 리다이렉트
+            return "redirect:/login";
+        }
+        Integer idUser = user.getIdUser();
+
         if (myPageService.isStudent(idUser)) {
             myPageService.deleteStudent(idUser);
             return "sample2";
@@ -97,33 +116,4 @@ public class MyPageController {
             return "/";
         }
     }
-
-
-//    @GetMapping("/stuMyPage")
-//    public String stuMyPage(Model model, @RequestParam Integer idUser) {
-//        model.addAttribute("user", myPageService.getStudentById(idUser));
-//        return "/myPage/stuMyPage";
-//    }
-//
-//    @PostMapping("/updateStuInfo")
-//    public String updateStuInfo(Student student) {
-//        myPageService.updateStudent(student);
-//
-//        return "redirect:/user/stuMyPage?idUser=" + student.getIdUser();
-//    }
-
-
-    ////myPage/profMyPage?idUser=1
-//@GetMapping("/profMyPage")
-//public String profMyPage(Model model, int idUser) {
-//    model.addAttribute("user", myPageService.getProfessorById(idUser));
-//    return "/myPage/profMyPage";
-//}
-//
-////myPage/adminMyPage?idUser=1234
-//@GetMapping("/adminMyPage")
-//public String adminMyPage(Model model, int idUser) {
-//    model.addAttribute("user", myPageService.getAdminById(idUser));
-//    return "/myPage/adminMyPage";
-//}
 }
